Move InstallEndpoint to its own file
